// backend/src/models/types.ts
export interface Geo {
  lat: string;
  lng: string;
}

export interface Address {
  street: string;
  suite: string;
  city: string;
  zipcode: string;
  geo: Geo;
}

export interface Company {
  name: string;
  catchPhrase: string;
  bs: string;
}

export interface User {
  id: number;
  name: string;
  username: string;
  email: string;
  address: Address;
  phone: string;
  website: string;
  company: Company;
}

export interface Post {
  userId: number;
  id: number;
  title: string;
  body: string;
}

export interface Comment {
  postId: number;
  id: number;
  name: string;
  email: string;
  body: string;
}

export interface UserWithPosts extends User {
  posts: PostWithComments[];
}

export interface PostWithComments extends Post {
  comments: Comment[];
}

export interface PaginatedResponse<T> {
  total: number;
  page: number;
  pages: number;
  limit: number;
  data: T[];
}